Fix AddMemberBySBS ordering
Sweeps should happen first and adding member second, because that's the
order in which team signatures will be posted.

Also changed some tx internal calls that should never return an error to
not return an error. All calls that actually create or modify payloads
should never fail if preconditions are satisfied, so it's possible to do
all checks beforehand and then be able to modify AddMemberTx object with
success guaranteed.